add main func
