+++ app.js
@@ -9,11 +9,12 @@ const cookieParser = require('cookie-parser');
 const cors = require('cors');
 const helmet = require('helmet');
 const corsOption = require('./utils/corsOption');
-const { createUser, login } = require('./controllers/controllers');
+const { createUser, login } = require('./controllers/users');
 const auth = require('./middlewares/auth');
 const NotFoundError = require('./errors/NotFoundError');
 const errHandler = require('./middlewares/errHandler');
-const router = require('./routes/routes');
+const usersRouter = require('./routes/users');
+const moviesRouter = require('./routes/movies');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 
 const { PORT = 3000 } = process.env;
@@ -33,7 +34,7 @@ app.post(
   '/signup',
   celebrate({
     body: Joi.object().keys({
-      name: Joi.string().min(2).max(30),
+      name: Joi.string().min(2).max(30).required(),
       email: Joi.string().email().required(),
       password: Joi.string().required(),
     }),
@@ -54,7 +55,8 @@ app.get('/signout', (_, res) => {
   res.status(200).clearCookie('jwt').send({ message: 'Выход' });
 });
 app.use(auth);
-app.use(router);
+app.use(moviesRouter);
+app.use(usersRouter);
 
 app.use('/*', (req, res, next) => {
   next(new NotFoundError('Страница по указанному URL не найдена'));
