+++ models/movie.js
@@ -64,7 +64,6 @@ const MovieSchema = new mongoose.Schema(
       type: String,
       required: true,
     },
-
   },
   {
     versionKey: false,
